Release job version verification checks for an exact match
The step in the release process that checks whether a release number has
already been used, we are failing while trying to ship v2.2.2 because
v2.2.2rc1 already exists. We need to check for an exact match rather than
doing a simple grep match.

Use `git rev-parse` to do an exact match check for whether a git tag
already exists.