Commit changes to version number in release workflow (#182)
## Problem

The version number bump was not automatically committed after the v2.2.2
release. This was the correct behavior for prerelease versions (which is
mainly how I tested the workflow prior to landing it), but is not
correct for prod releases.

## Solution

Add a step in the workflow to commit version number changes when
releasing.

## Type of Change

- [x] Infrastructure change (CI configs, etc)

## Test Plan

No good way to test this, really, until we are ready to do the next
release.