Add pdoc integration in GitHub workflows (#207)
## Problem
We're working on generating documentation references from the codebase
as a part of CI to each client.

For Python, we can use [pdoc](https://pdoc.dev/) to generate our client
reference as a static webpage which can then be directly deployed to
GitHub pages.

We also need to update CI workflows to handle building and deploying
documentation artifacts.

## Solution
- Add `pdoc` as a dependency under
`[tool.poetry.group.dev.dependencies]` in `pyproject.toml`.
- Add a new `build-docs` action under `.github/actions/`.
- Add a `build-docs` job to `pr.yaml` to verify docs build along with
testing on prs.
- Add new `merge` workflow under `.github/workflows/`.
- Update some of the existing docstrings in `index.py` and `manage.py`.

**Note:** There's a lot more that needs to be cleaned up with the
docstrings and overall formatting of the `pdoc` output. I felt like the
PR was getting large so opted to keep this one focused on the CI pieces
and getting things properly building in the pipeline. I'll be following
up with more fine-grained comment coverage of the top-level modules.

I have my default vscode formatter set to black, and I think prettier
went a little wild with some of the yaml files. I can try and back out
the changes but most of them seem fine to me, although slightly
different from some of the styling after black was run the other day.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)
- [X] Infrastructure change (CI configs, etc)
- [X] Non-code change (docs, etc)

## Test Plan
Make sure the `Pull Request / Build docs with pdoc` check passes as a
part of this PR.

I forked the repo and validated the `merge` workflow ran as expected and
deployed the docs properly.
You can see that run here:
https://github.com/austin-denoble/pinecone-python-client/actions/runs/6359304970

You can see the deployed docs in GitHub pages for validation of the
overall flow here:
https://austin-denoble.github.io/pinecone-python-client/pinecone.html