Support passing custom host for Index operations (#218)
## Problem
We need support for supplying a custom `host` value for index
operations.

## Solution
`pinecone.init()` and `Config` already support the ability to supply a
custom `host`. It didn't seem like it was being passed through to the
`ApiClient`:

- Update `_get_api_instance()` in `manage.py` to check for a
`Config.CONTROLLER_HOST` value, and apply it to the `client_config` if
so. This will then be used throughout the index operations.
- Add a unit test file for `manage.py`, test setting `controller_host`
and timeout functionality in `create_index` / `delete_index`.

## Type of Change
- [X] New feature (non-breaking change which adds functionality)

## Test Plan
Pull this branch down and test locally. I validated myself by passing a
bogus `host` in and looking over the errors, then passing the proper
`host` via `init` to validate it works properly.

```python
>>> poetry shell
>>> python3

>>> import pinecone
>>> pinecone.init(api_key="123-456-789", environment="my_environment", host: "https:custom-host-foo.com")

# Having passed a custom host that's inaccessible should fail
>>> pinecone.list_indexes()
# urllib3.exceptions.MaxRetryError: HTTPSConnectionPool(host='fooblahblah.io', port=443): Max retries exceeded with  
# url: /databases (Caused by NameResolutionError("<urllib3.connection.HTTPSConnection object at 0x107eb9dd0>:
# Failed to resolve 'fooblahblah.io' ([Errno 8] nodename nor servname provided, or not known)"))


# Re-Init - should work as expected
>>> pinecone.init(api_key="123-456-789", environment="my_environment")

pinecone.list_indexes()
# ['austin-3', 'austin-py-test', 'resin--test-index-1', 'test-create-1']
```