Downgrade to openapi 5.2, add template overrides (#229)
## Problem

Recent versions of openapi generator have added dependencies we don't
want to take on, especially `pydantic` v1. Pydantic v2 has been out for
a while and we think this is very likely to lead to dependency clashes
for some users.

## Solution

The reason we updated in the first place was to avoid generating code
that emitted urllib3 deprecation notices. In this PR, I downgrade to the
previously used 5.2.0 generator version and address the urllib3 error
using [template
overrides](https://openapi-generator.tech/docs/templating/) in the
generation step.

In this diff:
- Go back to generating the core with openapi 5.2.0 (same version as
before, e.g. on main)
- Everything under `pinecone/core` is generated using openapi and don't
need to be specifically reviewed. This is like 90% of the files changed.
- Use a template override in the generation step (handled in a private
repo) to adjust the urllib3 getheaders() usage and avoid spamming those
deprecation notices.
- Notice that generated `rest.py` is using [correct
syntax](https://stackoverflow.com/questions/75382508/deprecationwarning-httpresponse-getheader-is-deprecated-and-will-be-removed-i)
for getting headers in urllib3
- Fix tests that broke along the way, mostly due to changes in package
structure used by the generated code.
- Remove dependencies from `pyproject.toml` that were added when
previously going from openapi 5.2.0 => 7.0.1 (pydantic, aenum)


## Type of Change

- [x] None of the above: Update generated code, change dependencies

## Follow-up needed

- Docs build broken. By the looks of it, it seems related to numpy. But
I want to remove numpy in a follow-up diff anyway so I may merge with
this build failing for now.