[spruce] Pin docs-build python version (#235)
## Problem

This is a cherry-pick of changes in #234 onto the spruce branch. I want
this change in both `main` and `spruce` but rebasing spruce or merging
main into spruce is complicated due to the number of open PR branches on
spruce at the moment.

## Solution

Describe the approach you took. Link to any relevant bugs, issues, docs,
or other resources.

## Type of Change

- [x] Infrastructure change (CI configs, etc)