[spruce] Fix issue with polling after delete index (#253)
## Problem

There was a small error in a recent change to add a `names()` utility
method on the response from `Pinecone#list_indexes` to facilitate easy
iteration over index names. This resulted in an error when deleting
indexes:

```
    def get_remaining():
>       return name in api_instance.list_indexes().names()
E       TypeError: 'NoneType' object is not callable
```

This error got introduced because `api_instance.list_indexes()` and
`self.list_indexes()` are both a thing in this context, but only
`self.list_indexes()` returns a result with the `.names()` helper.

## Solution

Call names on the correct object to avoid the error.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

I confirmed with manual testing that this resolves the issue.