[Bug] Stop coercion of strings into datetime objects (#255)
## Problem

Sometimes when strings are stored as metadata values, later fetching or
querying causes those values to be coerced into datetime objects by the
client before they are presented to the caller. This is a default
behavior in the generated openapi client that is a result of the python
`datetime` constructor working quite hard to turn things into dates.
We've seen scenarios where this was not desired, so for now we will
disable this conversion process to have consistent handling of all
strings. People who desire date objects can easily implement the
conversion themselves.

## Solution

- Update openapi templates to comment out and remove logic related to
converting strings into date and datetime objects.
- Regenerate client with new templates
- Add unit test verifying the type coercion does not take place. 

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

`poetry run pytest tests/unit/data`