Remove deprecated `queries` option in `query()` (#272)
## Problem

There was an argument on the `query()` method for passing multiple
queries at once. This has long been deprecated in the API and SDK, so
it's time to remove it.
 
## Solution

- Clean up the `queries` option to both `Index#query` and
`IndexGRPC#query`.
- Remove code no longer needed to interpret the many forms these query
options could take
- Cleanup and simplify response parsing logic
- Remove related tests passing `queries` kwarg

## Type of Change

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Test Plan

Recently added integration test suite gives high confidence queries
still work after this change.