Update configure_index index for new API (#281)
## Problem

The shape of the configure_index shape has changed, so we need to pass
different stuff.

## Solution

- Construct new request body
- Add minimal integration test. I just want to see the request succeed
for now. The time needed for changes to take effect is variable, and I'm
in a bit of a rush. So I don't want to do a bunch of polling for status
updates in my test to verify the server actually did the correct thing.
- Delete some other legacy tests in `integ/` that have been disabled for
a while now.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

New test added.