Adjust supported python versions to ^3.8 (#312)
## Problem

Extend test coverage for python 3.11 and 3.12. Adjust pyproject.toml
python dependency to `^3.8` to ease some installation problems in
projects which depend on us.

## Solution

Tweak test matrix to add coverage for 3.11 and 3.12. Adjust
pyproject.toml to reflect specific version deps for grpcio and urllib3.

## Type of Change

- [x] New feature (non-breaking change which adds functionality)
- [x] Infrastructure change (CI configs, etc)

## Test Plan

Describe specific steps for validating this change.