Upgrade protobuf from v3.20.0 to v4.25.3
Upgrade protobuf from v3 to v4. This adds a number of performance
improvements in parsing / serialization as documented at
https://protobuf.dev/news/2022-05-06/#python-updates

This increases upsert() throughput by 3x (measured by upserting 1M 768
dimension indexes to a pod-based index in batches of 500):

Before:  880 vectors/sec
After:  2580 vectors/sec

As per the documentation, this results in an incompatible change with
the _generated_ Python code, so this depends on a related change to
pinecone-protos to change the version of protobuf used to generate the
Python code there.

This patch:

- Changes the version of protobuf package from v3 to current latest v4
  (v4.25.3).

- Updates the generated pb2 files from pinecone-protos to the protobuf
  v4 generated version.

- Switches from grpc-gateway-protoc-gen-openapiv2 to
  protoc-gen-openapiv2 as the support library needed to pull in the
  autogeneated openapi annotations dependancy - which changes format
  between v3 and v4.

- protobuf v4 returns less specific exception messages during
  serialisation in some cases, for those which we are testing for add
  explicit checks to return simialr messages as v3 did.