Support proxy_url and ssl_ca_certs options for gRPC
The PineconeGRPC client does not currently honor the proxy_url and
ssl_ca_certs arguments - they are accepted but silently ignored.

This means the gRPC client cannot be used in enviroments which require
a proxy to access the Pinecone service.

Address by wiring up these two config options to the corresponding
options in the underlying grpc library - in addition to them still
being passed to the HTTP-based control-plane connections.

Usage is very similar to the non-gRPC client - for example:

    pc = PineconeGRPC(my_api_key,
                      proxy_url="http://localhost:8080",
                      ssl_ca_certs="/path/to/my/ca-cert.pem")