do not override port if defined in host (#362)
## Problem

For an Index gRPC connection, port 443 is blindly added to the
configured host for setting up the endpoint. A user should be able to
specify a port for connection.

## Solution

This PR will no longer append ":443" to the configured gRPC endpoint if
a colon is already present in the host configuration.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Unit tests were updated for omitting and including a port in the host
config. Verified that the client could connect to an Index on a port
other than 443.