Rename package from pinecone-client to pinecone (#383)
## Problem

We want to rename this package from `pinecone-client` to `pinecone`.

## Solution

- Update `pyproject.toml` to refer to the new name `pinecone`
- Find and replace references to old package name in README and
docstrings
- Remove nightly publish job that would need to be updated but has been
broken for a long time for other reasons

After this rename, we will need to setup a separate project to wrap this
and continue publishing it under the old `pinecone-client` name for
people who may not know about the rename, very similar to how old
releases have been republished as `pinecone` using [this wrapper
project](https://github.com/pinecone-io/pypi-backfill-pinecone).

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [x] None of the above: We plan to publish a wrapper package under the
name `pinecone-client`. So this shouldn't be a breaking change for end
users consuming updates via the old name.

## Test Plan

Describe specific steps for validating this change.