[Fix] Add missing dateutil dep (#391)
## Problem

SDK version 5.3.0 was accidentally shipped with a missing dependency.
This wasn't caught in testing because the dependency was present in our
`poetry.lock`, which is used when installing in CI, but not in
`pyproject.toml` so it wasn't included for the published package. We got
into this state because we used to have a dependency on
`python-dateutil` which was removed and later re-added when date parsing
was needed for support of import endpoints (added 5.3.0).

Additional manual testing in google colab didn't catch this either
because that environment loads `python-dateutil` as part of the default
environment. 🤦‍♀️

## Solution

Add missing dependency.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)