Disable tests for weird ids string encoding (#412)
## Problem

We have a very large test matrix to check that different kinds of
strings are handled properly. But we probably don't get that much value
out of running these for every single PR.

## Solution

Gate these tests on a boolean so I can only run these when desired
instead of on every PR. It's most relevant to run them when changing
dependencies or python versions.

## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [x] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

Describe specific steps for validating this change.